feat(kbar): embed snippet/signature/task create forms in palette
Render "Create Snippet", "Create Signature", and "Create Task" as
in-chrome command-palette pages (matching the entity-instance create
flow) instead of popping separate modals.

- Extract shell-free SignatureForm / TaskForm cores; signature-dialog
  and task-dialog become thin Dialog wrappers (public API unchanged).
  TaskForm keeps its "Create more" toggle and exports the @mention-picker
  Esc carve-out, which the palette shell now applies on the task page.
- Snippet reuses the already shell-free SnippetForm directly.
- Add create-snippet / create-signature / create-task palette pages,
  store openers + carried params, and breadcrumb routing.
- Record-actions "Create task" now drills into the embedded task page
  with the record pre-linked instead of opening the standalone modal.

Standalone flows (settings pages, t,n / c,* hotkeys, sidebar, record
drawer) keep working through the existing dialog wrappers + stores.
