+++ apps/web/src/components/kbar/pages/create-signature.tsx
@@ -1,0 +1,27 @@
+// apps/web/src/components/kbar/pages/create-signature.tsx
+'use client'
+
+import { SignatureForm } from '~/components/signatures/ui/signature-form'
+import { useCommandPaletteStore } from '../store'
+
+/**
+ * Hosts the shell-free {@link SignatureForm} as a palette page (create mode only).
+ * The breadcrumb supplies the title (no header slot). On success / cancel the
+ * palette closes; back returns to root.
+ */
+export function CreateSignaturePage() {
+  const page = useCommandPaletteStore((s) => s.page)
+  const close = useCommandPaletteStore((s) => s.close)
+  const goTo = useCommandPaletteStore((s) => s.goTo)
+
+  return (
+    <div className='p-4'>
+      <SignatureForm
+        open={page === 'create-signature'}
+        onSuccess={close}
+        onClose={close}
+        onCancel={() => goTo('root')}
+      />
+    </div>
+  )
+}
